Gemini_RAG_pipline/requirements.txt
# Core framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Env vars & logging
python-dotenv==1.0.1
loguru==0.7.2

# Database (pick what you need)
# MongoDB
motor==3.3.2
pymongo==4.6.1

# SQL Databases (Install only if needed)
sqlalchemy==2.0.34
pymysql==1.1.1
psycopg2-binary==2.9.9

# AI / ML Dependencies
google-generativeai==0.8.3
sentence-transformers==3.0.1
faiss-cpu==1.8.0

# Utilities
requests==2.32.3
pandas==2.2.3

# Document Processing
python-docx==1.1.0
PyPDF2==3.0.1



